# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "STRNG API",
    "version": "1.0.0",
    "description": "API REST de la plataforma STRNG para catálogo, pedidos, envíos y repartidores."
  },
  "servers": [{ "url": "http://localhost:3001/api", "description": "Desarrollo local" }],
...
  "paths": {
...
    "/auth/login": { "post": { "tags": ["Auth"], "summary": "Iniciar sesión", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } } } }, "responses": { "200": { "description": "Tokens emitidos", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" } } } },
    "/auth/refresh": { "post": { "tags": ["Auth"], "summary": "Rotar refresh token", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RefreshRequest" } } } }, "responses": { "200": { "$ref": "#/components/responses/AuthResponse" }, "401": { "$ref": "#/components/responses/Unauthorized" } } } },
    "/auth/verify": { "get": { "tags": ["Auth"], "security": [{ "bearerAuth": [] }], "summary": "Verificar access token", "responses": { "200": { "description": "Token válido" }, "401": { "$ref": "#/components/responses/Unauthorized" } } } },
...
  },
  "components": {
    "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT", "description": "Access token JWT de corta duración. Los scopes se entregan en la respuesta de login y se validan en el backend." }, "oauth2": { "type": "oauth2", "flows": { "password": { "tokenUrl": "/auth/login", "refreshUrl": "/auth/refresh", "scopes": { "productos:read": "Consultar catálogo", "productos:write": "Gestionar productos", "paquetes:read": "Consultar pedidos", "paquetes:write": "Gestionar pedidos", "repartidores:read": "Consultar repartidores", "repartidores:write": "Gestionar repartidores" } } } } },
...
    "responses": { "BadRequest": { "description": "Solicitud inválida" }, "Unauthorized": { "description": "Autenticación requerida" }, "Forbidden": { "description": "Permisos insuficientes" }, "NotFound": { "description": "Recurso no encontrado" }, "AuthResponse": { "description": "Tokens de sesión", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } } } } },
    "schemas": {
      "LoginRequest": { "type": "object", "required": ["username", "password"], "properties": { "username": { "type": "string" }, "password": { "type": "string", "format": "password" } } },
      "RefreshRequest": { "type": "object", "required": ["refreshToken"], "properties": { "refreshToken": { "type": "string" } } },
      "AuthResponse": { "type": "object", "properties": { "ok": { "type": "boolean" }, "accessToken": { "type": "string" }, "refreshToken": { "type": "string" }, "expiresIn": { "type": "integer" }, "user": { "$ref": "#/components/schemas/User" } } },
      "User": { "type": "object", "properties": { "id": { "type": "string" }, "username": { "type": "string" }, "role": { "type": "string" }, "scopes": { "type": "array", "items": { "type": "string" } } } },
...
    }
  }
}